+++ snoozy-core/src/main/java/fi/jubic/snoozy/ServerConfiguration.java
@@ -1,11 +1,35 @@
 package fi.jubic.snoozy;
 
+import fi.jubic.snoozy.swagger.SwaggerConfig;
+
 public interface ServerConfiguration {
+    /**
+     * Server bind address.
+     */
     String getHostname();
 
+    /**
+     * Server listen port.
+     */
     int getPort();
 
+    default boolean isDevMode() {
+        return false;
+    }
+
     default MultipartConfig getMultipartConfig() {
-        return new MultipartConfig() {};
+        return new DefaultMultipartConfig();
+    }
+
+    default SwaggerConfig getSwaggerConfig() {
+        return new DefaultSwaggerConfig();
+    }
+
+    class DefaultMultipartConfig implements MultipartConfig {
+
+    }
+
+    class DefaultSwaggerConfig implements SwaggerConfig {
+
     }
 }